0030047: [OEM - CoGen] New Animations for Credits
Reverted a fix which forces center-justification for Rolling animation. This is for compatibility reasons with Reveal 9 users. Another reason is that Title doesn't have Rolling Left/Right animations. If the style default horiz alignment is right or left, it won't be nice to change it to center with Rolling animation in Title shows.