0030950: [OEM - CoGen] Change default background for Scrapbook style to Victorian
Merged from Arcadia_BR3_branch rev 21766.

This updates the style's default theme and preview video to Victorian. It also updates the Japanese style description as provided by CoGen.

Note that this style is released to CoGen only. It is no longer bundled with muvee Reveal. We are now bundling Lifebook and Journal, which are essentially two split styles from Scrapbook.