#0032192: mes support for video rotater
We are no using Ultra Plain anymore but are instead using a custom Show*. 
I have commented out the Style enumeration. 
But for some reason that I need to investigate, at least one style needs to be loaded in MVCore to prevent an uncaught exception. 
Title and Credits code path has been commented.