#0026194: General code cleanup
Replacing the assert:

MVASSERT( startTime > (m_showTime+.1) , 
	  "Potential Audio Breakup. startTime = %f secs, m_showTime = %f secs, lead = %f secs", 
	  startTime, m_showTime, startTime - m_showTime );  


with an MVWARN which has the same parameters as the above.

why ah? Well, the assert would confuse mufat and technically, the potential breakup should rightly be a warning because it's not a software bug. The user essentially has a crappy grahics card, which he should dump and a buy a better one.