0032345: New construction mode for Cinnamon
Reset cutHintsEarlyShiftTime_secs to zero. Don't know why Kumar shift it earlier by 100ms in rev 16326. But I find that it is more important to be consistent with the descriptor output times (i.e. no timing offsets). Moreover, styles have been assuming that the cut hint times are exact.

Removed the necessity to force 10ms audio fade duration, even when fade is not set. This prevents unwanted "audio stutter" when concatenating music in an anchor-based production.

Made some changes to the wordings of the debug log and SegDump outputs, and removing redundant ones, making it easier to follow the construction logs and identify concatenated sections more quickly.